% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{calculate}
\alias{calculate}
\title{Calculate posterior probabilities in a BN on agena.ai cloud}
\usage{
calculate(input_model, login, dataSet = NULL, debug = FALSE)
}
\arguments{
\item{input_model}{an R model object}

\item{login}{an agena.ai cloud login}

\item{dataSet}{a dataSet in the R model object}

\item{debug}{boolean parameter to display debug messages or not}
}
\value{
BN inference results in the model
}
\description{
A function to send an input Bayesian network model to Agena AI Cloud servers.
Once called, the function will check authentication status, if it has not expired,
it will send the POST request with the model to the servers, and receive calculation
results to update the Bayesian network model (filling the results field with calculation results).
}
