% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{create_csv_template}
\alias{create_csv_template}
\title{Create a CSV template for a model for batch observations}
\usage{
create_csv_template(inputModel)
}
\arguments{
\item{inputModel}{an R model object}
}
\value{
csv file
}
\description{
This function creates an empty CSV file with the correct format so that it can
be filled in and used for create_batch_bases.
}
