% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{dbClose}
\alias{dbClose}
\title{dbClose}
\usage{
dbClose(conn)
}
\arguments{
\item{conn}{The connector object.}
}
\description{
Method of closing connection with DolphinDB server.
}
\examples{
dbClose(conn)
}
