% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{dbConnect}
\alias{dbConnect}
\title{dbConnect}
\usage{
dbConnect(conn, host, port, username = NULL, password = NULL)
}
\arguments{
\item{conn}{The connector object.}

\item{host}{The host running the DolphinDB server.}

\item{port}{The port running the DolphinDB server.}

\item{username}{The username to login onto server. (NULL if default)}

\item{password}{The password to login onto server. (NULL if default)}
}
\value{
Return the connector object with status of connection.
}
\description{
Method of getting connection with DolphinDB server.

(If the input contains 'username' and 'password', a log-in job will be done)

Currently, there is at most one TCP connection between the R API and dolphindb. If you call dbConnect multiple times, it closes the previously created connection and creates a new one
}
\examples{
conn <- dbConnect(DolphinDB(), "localhost", 8848)    # Without log-in
# conn <- dbConnect(DolphinDB(), "localhost", 8848, "username", "password")

if (conn@connected == TRUE) {
    # TO DO ...
    # dbRun(...)
}
}
