% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{dbRpc}
\alias{dbRpc}
\title{dbRpc}
\usage{
dbRpc(conn, func, args)
}
\arguments{
\item{conn}{The connector object.}

\item{func}{The function name running on the DolphinDB server.}

\item{args}{The list containing all arguments needed to be uploaded to server.}
}
\value{
Return the excuted result of function from server in R type.
}
\description{
Method of running remote procedure call.
}
\examples{
result <- dbRpc(conn, "concat", list("hello", "world"))
print(result)
}
