% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{dbRun}
\alias{dbRun}
\title{dbRun}
\usage{
dbRun(conn, script)
}
\arguments{
\item{conn}{The connector object.}

\item{script}{The script running the DolphinDB server.}
}
\value{
Return the excuted result of script from server in R type.
}
\description{
Method of running scripts on DolphinDB.
}
\examples{
result <- dbRun(conn, "matrix(1.5 2.5 3.5, 4 NULL 6)")
print(result)
}
