% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{dbUpload}
\alias{dbUpload}
\title{dbUpload}
\usage{
dbUpload(conn, keys, args)
}
\arguments{
\item{conn}{The connector object.}

\item{keys}{The variable names. (Cannot be repeated)}

\item{args}{The list containing all arguments needed to be uploaded to server.}
}
\value{
Return the excuted result of uploading.
}
\description{
Method of uploading arguments to DolphinDB.
}
\examples{
result <- dbUpload(conn, c("x", "y"), list(2.5L, "hello"))
if (result == TRUE) {
    r_x <- dbRun(conn, "x")
    r_y <- dbRun(conn, "y")
    print(r_x)
    print(r_y)   
}
}
