% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{DolphinDB}
\alias{DolphinDB}
\title{DolphinDB}
\usage{
DolphinDB()
}
\description{
Method to getting an object of connector.
}
\examples{
conn <- dbConnect(DolphinDB(), "localhost", 8848)
if (conn@connected == TRUE) {
    # TO DO ...
    # dbRun(...)
}
}
