% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{generate_cmpx}
\alias{generate_cmpx}
\title{Generate a CMPX file from R Model}
\usage{
generate_cmpx(inputModel)
}
\arguments{
\item{inputModel}{an R model object}
}
\value{
cmpx file
}
\description{
This function generates a .cmpx file for an agena.ai model
based on an R model object
}
