% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{GetPlacesIn}
\alias{GetPlacesIn}
\title{Return places of a specified type contained in specified places}
\usage{
GetPlacesIn(dcids, placeType)
}
\arguments{
\item{dcids}{required, dcid(s) identifying a containing place.
This parameter will accept a vector of strings
or a single-column tibble/data frame of strings.
To select a single column, use \code{select(df, col)}.}

\item{placeType}{required, string identifying the type of place to query for.}
}
\value{
Named list or column of places contained in each given dcid of the
  given placeType. If dcids input is vector of strings, will return a named
  list. If dcids input is tibble/data frame, will return a new single-column
  tibble/data frame.
}
\description{
Returns a mapping between the specified places
  and the places of a specified type contained in them.
}
\details{
Assigning output to a tibble/data frame will yield a list of contained
places. To convert this to 1-to-1 mapping (the containing place will
be repeated), use \code{tidyr::unnest}.
}
\examples{
# Atomic vector of the dcids of Santa Clara and Montgomery County.
countyDcids <- c('geoId/06085', 'geoId/24031')
# Get towns in Santa Clara and Montgomery County.
towns <- GetPlacesIn(countyDcids, 'Town')

# Tibble of the dcids of Santa Clara and Montgomery County.
df <- tibble(countyDcid = c('geoId/06085', 'geoId/24031'))
# Get towns in Santa Clara and Montgomery County.
df$townDcid <- GetPlacesIn(df, 'Town')
# Since GetPlacesIn returned a mapping between counties and
# a list of towns, use you can use tidyr::unnest to create
# a 1-1 mapping between each county and its towns.
}
