% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popobs.R
\name{GetPopulations}
\alias{GetPopulations}
\title{Return StatisticalPopulations associated with specified dcids}
\usage{
GetPopulations(dcids, populationType, constraintsPVMap)
}
\arguments{
\item{dcids}{required, dcid(s) identifying Places of populations to query
for. This parameter will accept a vector of strings, or a single column
tibble/data frame of strings. To select a single column from multicol
tibble/data frame, use \code{select(df, col)}.}

\item{populationType}{required, string identifying the
population type of the StatisticalPopulation.}

\item{constraintsPVMap}{required, named list of constraining properties and
desired values that defines the population to get. Example:
\code{list(employment = 'BLS_Employed')}}
}
\value{
dcid strings of populations corresponding to each given Place
  satisfying the constraints.
  Will be encapsulated in a named list if dcids input is vector of strings,
  or a new single column tibble if dcids input is tibble/data frame.
}
\description{
Returns a mapping between the each given Place and the
  StatisticalPopulation satisfying all specified constraints
  contained in the Place.
}
\examples{
# Set the dcid to be that of Santa Clara County.
sccDcid <- 'geoId/06085'
# Get the population dcids for Santa Clara County.
malePops <- GetPopulations(sccDcid, 'Person', list(gender = 'Male'))
femalePops <- GetPopulations(sccDcid, 'Person', list(gender = 'Female'))

# Create an atomic vector of the dcids of California, Kentucky and Maryland.
stateDcids <- c('geoId/06', 'geoId/21', 'geoId/24')
# Get the population dcids for each state.
malePops <- GetPopulations(stateDcids, 'Person', list(gender = 'Male'))
femalePops <- GetPopulations(stateDcids, 'Person', list(gender = 'Female'))

# Create an tibble of the dcids of California, Kentucky, and Maryland.
df <- tibble(countyDcid = c('geoId/06', 'geoId/21', 'geoId/24'),
             rand = c(1, 2, 3))
# Get the population dcids for each state.
df$malePops <- GetPopulations(select(df, countyDcid), 'Person',
                              list(gender = 'Male'))
df$femalePops <- GetPopulations(select(df, countyDcid), 'Person',
                                list(gender = 'Female'))
}
