% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-core.R
\name{GetPropertyLabels}
\alias{GetPropertyLabels}
\title{Return property labels of specified nodes}
\usage{
GetPropertyLabels(dcids, outgoing = TRUE)
}
\arguments{
\item{dcids}{required, vector of string(s) that identify node(s) to get
property labels for.}

\item{outgoing}{optional, boolean indicating whether to get properties
originating from
the given node. TRUE by default.}
}
\value{
Named list of properties associated with the given dcid(s) via the
  given direction.
}
\description{
Returns a map between nodes and outgoing (default) or incoming
  property labels.
}
\examples{
# dcid string of Santa Clara County.
sccDcid <- 'geoId/06085'
# Get incoming and outgoing properties for Santa Clara County.
inLabels <- GetPropertyLabels(sccDcid, outgoing = FALSE)
outLabels <- GetPropertyLabels(sccDcid)

# List of dcid strings of Florida, Planned Parenthood West, and the
# Republican Party.
dcids <- c('geoId/12', 'plannedParenthood-PlannedParenthoodWest',
           'politicalParty/RepublicanParty')
# Get incoming and outgoing properties for Santa Clara County.
inLabels <- GetPropertyLabels(dcids, outgoing = FALSE)
outLabels <- GetPropertyLabels(dcids)
}
