% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-core.R
\name{GetTriples}
\alias{GetTriples}
\title{Return all triples involving specified nodes}
\usage{
GetTriples(dcids, limit = 100)
}
\arguments{
\item{dcids}{required, vector of string(s) that uniquely identify
the node(s) to get triples for.}

\item{limit}{optional, integer indicating the max number of triples to get
for each property. 100 by default.}
}
\value{
Map between each dcid and all triples with the dcid as the subject or
object. Triples are represented as (subject, predicate, object).
}
\description{
Returns all triples (subject-predicate-object) where the specified node is
either a subject or an object.
}
\examples{
# Set the dcid to be that of Santa Clara County.
sccDcid <- 'geoId/06085'
# Get triples.
triples <- GetTriples(sccDcid)

# List of dcid strings of Florida, Planned Parenthood West, and the
# Republican Party.
dcids <- c('geoId/12', 'plannedParenthood-PlannedParenthoodWest',
           'politicalParty/RepublicanParty')
# Get triples.
triples <- GetPropertyLabels(dcids)
}
