% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{local_api_calculate}
\alias{local_api_calculate}
\title{Local agena.ai API model calculation}
\usage{
local_api_calculate(model, dataSet, output)
}
\arguments{
\item{model}{an R model object}

\item{dataSet}{a dataSet in the model}

\item{output}{file name for the output json - just the filename using the working directory, not a full file path}
}
\value{
a json file for the model with results
}
\description{
Local agena.ai API model calculation
}
