% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{local_api_compile}
\alias{local_api_compile}
\title{Local agena.ai API compilation}
\usage{
local_api_compile()
}
\value{
no return value, used to compile maven in the local API directory
}
\description{
sets the version and compiles the maven environment for
the local agena.ai developer API in the working directory
}
