% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{login}
\alias{login}
\title{Login to agena.ai cloud}
\usage{
login(username, password)
}
\arguments{
\item{username}{agena.ai cloud username}

\item{password}{agena.ai cloud password}
}
\value{
no return value, used in to authenticate user credentials
}
\description{
Login to agena.ai cloud
}
