% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_doc.R
\docType{package}
\name{RDolphinDB}
\alias{RDolphinDB}
\alias{RDolphinDB-package}
\title{RDolphinDB : Connecting to DolphinDB with R}
\description{
The R API of DolphinDB.

You can use R connecting to DolphinDB with IP address and port,

and then running scripts or functions and geting result in R form.
}
\section{RDolphinDB functions}{

DolphinDB()
dbConnect()
dbRun()
dbRpc()
dbClose()
}

\author{
Jingtang Zhang <jingtang.zhang@dolphindb.com>
}
