% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Sensitivity analysis on agena.ai cloud}
\usage{
sensitivity_analysis(input_model, login, sensitivity_config)
}
\arguments{
\item{input_model}{an R model object}

\item{login}{an agena.ai cloud login}

\item{sensitivity_config}{a sensitivity analysis config object}
}
\value{
sensitivity analysis report
}
\description{
Sensitivity analysis on agena.ai cloud
}
