% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{SetApiKey}
\alias{SetApiKey}
\title{Set your API Key}
\usage{
SetApiKey(key)
}
\arguments{
\item{key}{(string) your API Key.}
}
\description{
Using the Data Commons API requires you to provision an API key on
GCP. See
\url{https://datacommons.readthedocs.io/en/latest/started.html#creating-an-api-key}
}
\examples{
# Say you got an API Key: 12345678, in the R console:
SetApiKey("12345678")
}
