% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDolphinDB.R
\name{DolphinDB}
\alias{DolphinDB}
\title{DolphinDB}
\usage{
DolphinDB()
}
\description{
Method to getting an object of connector.
}
\examples{
conn <- DolphinDB()
conn <- dbConnect(conn, "localhost", 8848)

# Recommanded
conn <- dbConnect(DolphinDB(), "localhost", 8848)
}
