% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popobs.R
\name{GetObservations}
\alias{GetObservations}
\title{Return dcids of Observations observing the given dcids}
\usage{
GetObservations(dcids, measuredProperty, statsType, observationDate,
  observationPeriod = NULL, measurementMethod = NULL)
}
\arguments{
\item{dcids}{required, dcid(s) to get observations for. The returned
Observations must have observedNode
\url{https://browser.datacommons.org/kg?dcid=observedNode} as one of
these specified dcids.
This parameter will accept a vector of strings,
or a single column tibble/data frame of strings.
To select a single column from multicol tibble/data frame, use
\code{select(df, col)}.}

\item{measuredProperty}{required, string identifying the measured property.}

\item{statsType}{required, string identifying the statistical type of the
measurement.}

\item{observationDate}{required, string specifying the observation date in
ISO8601 format.}

\item{observationPeriod}{optional, string specifying the observation period.
E.g. 'P1Y': Period 1 year, 'P2M': Period 2 month, 'P3D': Period 3 day,
'P4m': period 4 minute.}

\item{measurementMethod}{optional, string specifying a
measurement method.}
}
\value{
Observation values corresponding to each given dcid.
  Will be encapsulated in a named list if dcids input is vector of strings
  or a new single column tibble if dcids input is tibble/data frame.

  If dcids input is a vector of strings, returns a named
  list mapping a given dcid to the unique Observation
  observing the dcid, where the observation is specified by what is given
  in the other parameters.
  A given dcid will NOT be a member of the dict if such an observation
  does not exist.

  If dcids input is a tibble/data frame, returns a new single column
  tibble, where the i-th entry corresponds to the Observation
  observing the given dcid as specified by the other parameters, if
  such exists. Otherwise, the cell is empty.
}
\description{
Returns a mapping between each given dcid
  and the Observations on the specified
  property satisfying all specified constraints.
}
\examples{
stateDcids <- c('geoId/06', 'geoId/21', 'geoId/24')
# Using vectors
femalePops <- GetPopulations(stateDcids, 'Person', list(gender = 'Female'))
# Atomic vector version
# \\code{unlist} converts lists to vectors
femaleCount <- GetObservations(unlist(femalePops), 'count',
                               'measured_value', '2016',
                               measurementMethod = 'CenusACS5yrSurvey')

# Tibble/data frame version
df <- tibble(countyDcid = stateDcids, rand = c(1, 2, 3))
df$malePops <- GetPopulations(select(df, countyDcid), 'Person',
                              list(gender = 'Male'))
df$maleCount <- GetObservations(select(df, malePops), 'count',
                                'measured_value', '2016',
                                measurementMethod = 'CenusACS5yrSurvey')
}
