% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-core.R
\name{GetTriples}
\alias{GetTriples}
\title{Return all triples involving specified nodes}
\usage{
GetTriples(dcids, limit = 100)
}
\arguments{
\item{dcids}{required, vector of string(s) of dcids to get triples for.}

\item{limit}{optional, integer indicating the max number of triples to
return. 100 by default.}
}
\value{
Named list mapping dcids to a list of triples (s, p, o) where s, p,
and o are strings and either the subject or object is the mapped dcid.
}
\description{
A knowledge graph can be described as a collection of "triples" which are
3-tuples that take the form (s, p, o). Here, s and o are nodes in the graph
called the subject and object respectively while p is the property label
of a directed edge from s to o (sometimes also called the predicate).
}
\examples{
# Set the dcid to be that of Santa Clara County.
sccDcid <- 'geoId/06085'
# Get triples.
triples <- GetTriples(sccDcid)

# List of dcid strings of Florida, Planned Parenthood West, and the
# Republican Party.
dcids <- c('geoId/12', 'plannedParenthood-PlannedParenthoodWest',
           'politicalParty/RepublicanParty')
# Get triples.
triples <- GetPropertyLabels(dcids)
}
